Modifications for improved data storing * Dispatch document: Updated to log dates and dispatch transit states. Added constructors for initialization options ++ DispatchService : Updated the dispatch queries i. To use drone specific max weight when loading +++ DroneController: Added GET and PUT mapping for drone retrieval and update ++++ DroneService : Updated with method using DroneRepository Query methods

+++ src/main/java/com/geek0/drones/model/Medication.java
@@ -11,14 +11,14 @@ public class Medication {
     @Id
     private String id;
     private String name;
-    private int weight;
+    private double weight;
     @Indexed(unique = true)
     private String code;
     private String image;
 
     public Medication(
             String name,
-            int weight,
+            double weight,
             String code,
             String image)
     {
